Create pb_StreamlabsSystem.py
Largely copied from the WR command.

Unfortunately only pulls the top 50-60 runs from each category combination at the moment...

Pulls Twitch API for game / title information, and searches speedrun.com API for a list of the categories for the game.  Then compares the category names against the title to find the category being run. Finally it runs through the list of runs searching for the ID speedrun.com uses for the runner stated in line 17.